Move hero section into it's own component
