+++ pages/index.js
@@ -1,53 +1,5 @@
 import Head from 'next/head';
-import {
-  FaBehance,
-  FaDribbble,
-  FaInstagram,
-  FaLinkedin,
-  FaTwitter,
-  FaVimeo,
-  FaYoutube
-} from 'react-icons/fa';
-
-function SocialMediaLink(props) {
-  return (
-    <a
-      {...props}
-      target="_blank"
-      className="hover:text-gray-300 hover:scale-125 transform transition duration-300"
-    >
-      {props.children}
-    </a>
-  );
-}
-
-function SocialMediaIcons() {
-  return (
-    <div className="flex text-3xl justify-between max-w-sm">
-      <SocialMediaLink href="https://twitter.com/imse_ty">
-        <FaTwitter />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://www.instagram.com/imse.ty">
-        <FaInstagram />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://www.behance.net/imsety">
-        <FaBehance />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://www.linkedin.com/in/imsety">
-        <FaLinkedin />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://dribbble.com/Imsety">
-        <FaDribbble />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://vimeo.com/imsety">
-        <FaVimeo />
-      </SocialMediaLink>
-      <SocialMediaLink href="https://www.youtube.com/channel/UCVptL5EXU-Rd8lvuQvzDK3A">
-        <FaYoutube />
-      </SocialMediaLink>
-    </div>
-  );
-}
+import Hero from '../components/hero';
 
 export default function Home() {
   return (
@@ -71,15 +23,7 @@ export default function Home() {
                 className="w-36"
               />
             </header>
-            <div>
-              <h1 className="font-bold text-3xl md:text-4xl mb-3">
-                Hey, my name is Imsety Taylor
-              </h1>
-              <p className="mb-6 md:text-lg">
-                I’m a Motion Designer based in Atlanta, GA.
-              </p>
-              <SocialMediaIcons />
-            </div>
+            <Hero />
             <footer>
               <p className="text-sm">
                 © 2021 Imsety Taylor. All rights reserved.
